+++ src/components/Tabs/index.ts
@@ -1,3 +1,3 @@
-// Tabs from copy-ui @ 2025-10-05.
+// Tabs from copy-ui @ 2025-10-16.
 
 export { Tabs } from "./Tabs";