+++ src/components/Tabs/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-10-05: Initial version.
+- 2025-10-16: Adjust its styles - improve spacing and hover states.